datalake-monitoring/virtualization-monitoring/scripts/requirements.txt
psycopg2-binary>=2.9.0
PyYAML>=6.0
Jinja2>=3.1.0
